@@ -1,10 +1,10 @@
 {
   "name": "node-aes-gcm",
-  "version": "0.2.4",
+  "version": "0.2.5",
   "description": "AES GCM module for node.js that takes advantage of GCM authentication",
   "main": "build/Release/node_aes_gcm.node",
   "engines": {
-    "node": ">= 0.12"
+    "node": ">= 6.0.0"
   },
   "scripts": {
     "test": "mocha -R spec"
@@ -15,8 +15,7 @@
   },
   "devDependencies": {
     "should": "13.x.x",
-    "buffertools": "2.1.x",
-    "mocha": "4.x.x"
+    "mocha": "8.x.x"
   },
   "keywords": [
     "AES",
@@ -34,6 +33,6 @@
   "gypfile": true,
   "readmeFilename": "README.md",
   "dependencies": {
-    "nan": "2.8.x"
+    "nan": "2.14.x"
   }
 }
